#1 [setting] : core ui 모듈 변경
