+++ core/ui/build.gradle.kts
@@ -38,7 +38,6 @@ android {
 
 dependencies {
 
-    implementation(libs.androidx.core.ktx)
-    implementation(libs.androidx.appcompat)
+    implementation(libs.bundles.androidx)
     implementation(libs.material)
 }
